wip feat: search with mixed column
