+++ console/packages/starwhale-core/src/widget/withWidgetDynamicProps.tsx
@@ -87,7 +87,7 @@ export default function withWidgetDynamicProps(WrappedWidgetRender: WidgetRender
         }, [tableConfig])
         const inViewport = useIsInViewport(myRef as any)
         const { recordInfo, recordQuery: query } = useQueryDatasetList(tableName, tableOptions, false)
-
+        console.log(tableOptions)
         const inViewLoadRef = useRef(false)
         const tableNameRef = useRef('')
